+++ main/aronet_device_client.h
@@ -64,8 +64,8 @@ typedef struct {
     char part_number[32];
     char name[64];
     char description[256];
-    uint32_t quantity;
-    uint32_t reorder_level;
+    float quantity;        // fractional for parts measured in meters, kg, etc. (1 decimal place)
+    float reorder_level;
     char unit[16];
 } aronet_part_t;
 
@@ -192,8 +192,8 @@ aronet_error_t aronet_get_jobs(const char *status, aronet_job_t *jobs, uint32_t 
  */
 aronet_error_t aronet_get_parts(aronet_part_t *parts, uint32_t max_parts, uint32_t *count);
 
-/** Apply a positive or negative manual stock adjustment to a part. */
-aronet_error_t aronet_adjust_part(uint32_t part_id, int32_t quantity_change, const char *reason);
+/** Apply a positive or negative manual stock adjustment to a part (1 decimal place). */
+aronet_error_t aronet_adjust_part(uint32_t part_id, float quantity_change, const char *reason);
 
 /**
  * Get single part by ID
